requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
coverage==6.5.0
decorator==5.1.1
Django==4.1.3
django-redis==5.2.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.2.0
hiredis==2.0.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.2
jsonschema==4.17.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
uritemplate==4.1.1
wcwidth==0.2.5
